WIP: Move readStoredCredentials out of util
