+++ src/services/readStoredCredentials.js
@@ -1,0 +1,5 @@
+import { credentialsSelector, getStoredState } from '../redux';
+
+const readStoredCredentials = () => credentialsSelector(getStoredState());
+
+export default readStoredCredentials;
