+++ src/util.js
@@ -18,7 +18,6 @@ import {
   rearg,
   zip,
 } from 'lodash';
-import { credentialsSelector, getStoredState } from './redux';
 
 const roundUpToNearest = (n, toValue) => Math.ceil(n / toValue) * toValue;
 
@@ -72,8 +71,6 @@ const isWeekday = negate(isWeekend);
 
 const shortDay = d => isNil(d) ? '' : format(d, 'YYYY-MM-DD');
 
-const readStoredCredentials = () => credentialsSelector(getStoredState());
-
 export {
   concatMerge,
   isAfterDay,
@@ -85,7 +82,6 @@ export {
   maxValueBy,
   minValueBy,
   pageableQuery,
-  readStoredCredentials,
   resolveCursors,
   roundUpToNearest,
   shortDay,
